requirements.txt
uvicorn[standard]==0.20.0
gunicorn==20.1.0
starlette==0.22.0
psycopg2-binary
